+++ crates/rspack_cacheable_test/tests/macro/manual_cacheable_dyn/mod.rs
@@ -1,0 +1,2 @@
+mod basic;
+mod with_generics;
