+++ crates/rspack_cacheable_test/tests/macro/mod.rs
@@ -3,4 +3,3 @@ mod cacheable_dyn;
 mod disable_cacheable;
 mod manual_cacheable;
 mod manual_cacheable_dyn;
-mod manual_cacheable_dyn_with_generics;
